backend/core/requirements.txt
 streamlit
openai
langchain
langchain-openai
langchain-community
langchain-google-firestore
langchain-chroma
python-dotenv
google-cloud-firestore
pydantic
typing-extensions
protobuf==3.20.3
pandas
numpy
black
pylint
watchdog
setuptools
python-multipart
fastapi
uvicorn
requests
